% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_tr.R
\name{add_tr}
\alias{add_tr}
\title{Assert that all table rows are nested in tr tags}
\usage{
add_tr(table.Node, header, body)
}
\arguments{
\item{table.Node}{the table node}

\item{header}{information passed from the header argument}

\item{body}{information passed from the body argument}
}
\value{
the table node
}
\description{
Assert that all table rows are nested in tr tags
}

