% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup_and_checks.R
\name{check_nested_table}
\alias{check_nested_table}
\title{Check if a table is inside the designated table}
\usage{
check_nested_table(Node)
}
\arguments{
\item{Node}{the XML node to be checked}
}
\description{
Check if a table is inside the designated table
}

