% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_type}
\alias{check_type}
\title{Assemble Check input values}
\usage{
check_type(doc, which)
}
\arguments{
\item{doc}{the HTML document which can be a file name or a URL or an already parsed HTMLInternalDocument, or an HTML node of class XMLInternalElementNode, or a character vector containing the HTML content to parse and process.}

\item{which}{a vector identifying which tables to return from within the document (when such was specified). Either a numeric vector for the tables' rank or a character vector specifiying an XPath for the tables.}
}
\value{
A list of table nodes.
}
\description{
Assemble Check input values
}

