% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean_colnames}
\alias{clean_colnames}
\title{Produces correct column names}
\usage{
clean_colnames(header.name.table = NULL, colNames = NULL)
}
\arguments{
\item{header.name.table}{the table node}

\item{colNames}{either a self-specificed character vector for the column names or a function used on header.name.table}
}
\value{
a character vector
}
\description{
Produces correct column names
}

