% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbody_header.R
\name{create_inbody}
\alias{create_inbody}
\title{Reshape in table header information into wide format}
\usage{
create_inbody(tab, table.Node, trindex, xpath)
}
\arguments{
\item{tab}{the table data frame}

\item{table.Node}{the table node}

\item{trindex}{the tr index of the inbody rows}

\item{xpath}{the xpath for the inbody rows}
}
\value{
the modified R data frame
}
\description{
Reshape in table header information into wide format
}

