% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{equal_length}
\alias{equal_length}
\title{Check if all list vectors have equal length}
\usage{
equal_length(x)
}
\arguments{
\item{x}{list of vectors}
}
\value{
logical
}
\description{
Check if all list vectors have equal length
}

