% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_and_checks.R
\name{eval_body}
\alias{eval_body}
\title{Evaluate and deparse the body argument}
\usage{
eval_body(arg)
}
\arguments{
\item{arg}{the body argument}
}
\description{
Evaluate and deparse the body argument
}

