% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_and_checks.R
\name{eval_header}
\alias{eval_header}
\title{Evaluate and deparse the header argument}
\usage{
eval_header(arg)
}
\arguments{
\item{arg}{the header information}
}
\value{
evaluated header info
}
\description{
Evaluate and deparse the header argument
}

