% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_xpath.R
\name{get_body_xpath}
\alias{get_body_xpath}
\title{Return body xpath}
\usage{
get_body_xpath(table.Node, body)
}
\arguments{
\item{table.Node}{the table node}

\item{body}{an information for the body rows}
}
\value{
a character vector of XPath statements
}
\description{
Return body xpath
}

