% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cell_element}
\alias{get_cell_element}
\title{Extracts cells elements}
\usage{
get_cell_element(cells, tag = "td | th", elFun = bodyFun,
  rm_escape = NULL)
}
\arguments{
\item{cells}{a list of cell nodes}

\item{tag}{a character vector that provides information used in the XPath expression to extract the correct elements}

\item{elFun}{elFun argument}

\item{rm_escape}{rm_escape argument}
}
\value{
A list of table nodes
}
\description{
Extracts cells elements
}

