% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cells}
\alias{get_cells}
\title{Extract body cell values}
\usage{
get_cells(table.Node, body)
}
\arguments{
\item{table.Node}{the table node}

\item{body}{body identifying information}
}
\value{
list of body information
}
\description{
Extract body cell values
}

