% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_colspans}
\alias{get_colspans}
\title{Extracts colspan elements}
\usage{
get_colspans(cells, tag = "td | th")
}
\arguments{
\item{cells}{a list of cell nodes}

\item{tag}{a character vector that provides information used in the XPath expression to extract the correct elements}
}
\value{
A list of column information from the cells
}
\description{
Extracts colspan elements
}

