% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_head.R
\name{get_header_xpath}
\alias{get_header_xpath}
\title{Assemble an XPath expression for the header}
\usage{
get_header_xpath(table.Node, header)
}
\arguments{
\item{table.Node}{the table node}

\item{header}{an information on the header rows}
}
\value{
a character vector for the XPath statement
}
\description{
Assemble an XPath expression for the header
}

