% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_rowspans}
\alias{get_rowspans}
\title{Extracts row elements}
\usage{
get_rowspans(cells, tag = "td | th")
}
\arguments{
\item{cells}{a list of cell nodes}

\item{tag}{a character vector that provides information used in the XPath expression to extract the correct elements}
}
\value{
A list of row information from the cells
}
\description{
Extracts row elements
}

