% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_rows.R
\name{get_trindex}
\alias{get_trindex}
\title{Return trindex given an XPath}
\usage{
get_trindex(xpath, table.Node)
}
\arguments{
\item{xpath}{XPath}

\item{table.Node}{the table node}
}
\description{
Return trindex given an XPath
}

