% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_xpath.R
\name{get_xpath}
\alias{get_xpath}
\title{Assemble XPath expressions for header and body}
\usage{
get_xpath(table.Node, header, body)
}
\arguments{
\item{table.Node}{the table node}

\item{header}{an information for the header rows}

\item{body}{an information for the body rows}
}
\value{
a character vector of XPath statements
}
\description{
Assemble XPath expressions for header and body
}

