% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has_tag}
\alias{has_tag}
\title{Assert a specific tag in an XML node}
\usage{
has_tag(table.Node, tag)
}
\arguments{
\item{table.Node}{the table node}

\item{tag}{a character string for the tag name to be matched}
}
\value{
logical value showing whether there is such a tag
}
\description{
Assert a specific tag in an XML node
}

