% Generated by roxygen2 (4.0.2): do not edit by hand
\name{htmltable}
\alias{htmltable}
\title{Robust methods for extracting structured information out of HTML tables}
\usage{
htmltable(doc, which = NULL, header = NULL, headerSep = " >> ",
  body = NULL, bodyFun = function(node) XML::xmlValue(node),
  headerFun = function(node) XML::xmlValue(node), rm_superscript = T,
  rm_escape = " ", rm_footnotes = T, colNames = NULL, ...)
}
\arguments{
\item{doc}{the HTML document which can be a file name or a URL or an already parsed document (by XML's parsing functions}

\item{which}{a vector identifying which tables to return from the document. Either a numeric vector for the tables' rank (no negative indexes allowed) or a character vector specifiying an XPath for the table}

\item{header}{a vector specifying header information. A numeric vector can be specified where each element corresponds to the table rows. A character vector may be specified that describes an XPath for the header rows. If left unspecified, htmltable tries to use semantic information from the HTML code}

\item{headerSep}{a character vector that is used as a seperator in the construction of the table's variable names (default value ' >> ')}

\item{body}{a vector that specifies which table rows should be used as body information. A numeric vector can be specified where each element corresponds to a table row (no negative indexes allowed). A character vector may be specified that describes an XPath for the body rows. If left unspecified, htmltable tries to use semantic information from the HTML code}

\item{bodyFun}{a function that is executed over the body cell nodes}

\item{headerFun}{a function that is executed over the header cell nodes}

\item{rm_superscript}{logical, denotes whether superscript information should be removed from header and body cells (default value TRUE)}

\item{rm_escape}{a character vector that, if specified,is used to replace escape sequences in header and body cells (default value ' ')}

\item{rm_footnotes}{logical, denotes whether semantic footer information should be removed (default value TRUE)}

\item{colNames}{a character vector of column names, or a function that can be used to replace specific column names (default value NULL)}

\item{...}{additional arguments}
}
\value{
An R data frame
}
\description{
Robust methods for extracting structured information out of HTML tables
}
\examples{
# When no spans are presented, htmltable produces output identical to XML's readHTMLTable()

url <- "http://en.wikipedia.org/wiki/World_population"
xp <- "//caption[text() = 'World historical and predicted populations (in millions)']/ancestor::table"
htmltable(doc = url, which = xp)

library(magrittr)
library(stringr)
library(XML)
popFun <- function(node) xmlValue(node) \%>\% str_replace(., ',', '')
htmltable(doc = url, which = xp, bodyFun = popFun)

#This table lacks header information. We provide them through colNames
doc <- "http://en.wikipedia.org/wiki/FC_Bayern_Munich"
xp2 <- "//td[text() = 'Head coach']/ancestor::table"
htmltable(doc = doc, which = xp2, encoding = "UTF-8", colNames = c("name", "role"))


#htmltable recognizes column spans and produces a one-dimension vector of variable information,
#also removes automatically superscript information since these are usually not wanted.

doc <- "http://en.wikipedia.org/wiki/Usage_share_of_web_browsers"
xp3 <-  "//table[5]"
bFun <- function(node) {xmlValue(node) \%>\% str_replace(., '\%$', '') \%>\% ifelse(equals(., ''), NA, .)}
htmltable(doc = doc, which = xp3, bodyFun = bFun)
}

