% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_rows.R
\name{identify_elements}
\alias{identify_elements}
\title{Assemble XPath expressions for header and body}
\usage{
identify_elements(table.Node, header, body, complementary = T)
}
\arguments{
\item{table.Node}{the table node}

\item{header}{a vector that contains information for the identification of the header row(s). A numeric vector can be specified where each element corresponds to the table rows. A character vector may be specified that describes an XPath for the header rows. If left unspecified, htmltable tries to use semantic information from the HTML code}

\item{body}{a vector that specifies which table rows should be used as body information. A numeric vector can be specified where each element corresponds to a table row. A character vector may be specified that describes an XPath for the body rows. If left unspecified, htmltable tries to use semantic information from the HTML code}

\item{complementary}{logical, should htmltab ensure complementarity of header, inbody header and
body elements (default TRUE)?}
}
\value{
a character vector of XPath statements
}
\description{
Assemble XPath expressions for header and body
}

