% Generated by roxygen2 (4.0.2): do not edit by hand
\name{left_value}
\alias{left_value}
\title{Fill cell with left-side neighbor value}
\usage{
left_value(table.Node)
}
\arguments{
\item{Node}{the HTML document which can be a file name or a URL or an already parsed HTMLInternalDocument, or an HTML node of class XMLInternalElementNode, or a character vector containing the HTML content to parse and process}

\item{tag}{a character string for the tag name to be matched}
}
\value{
logical
}
\description{
Fill cell with left-side neighbor value
}
\examples{
add(1, 1)
add(10, 1)
}

