% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_and_checks.R
\name{normalize_tr}
\alias{normalize_tr}
\title{Normalizes rows to be nested in tr tags, header in thead, body in tbody and numbers them}
\usage{
normalize_tr(table.Node)
}
\arguments{
\item{table.Node}{the table node}
}
\value{
the revised table node
}
\description{
Normalizes rows to be nested in tr tags, header in thead, body in tbody and numbers them
}

