% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_rows.R
\name{num_xpath}
\alias{num_xpath}
\title{num_xpath: Generate numeric XPath expression}
\usage{
num_xpath(data)
}
\arguments{
\item{data}{the header XPath}
}
\description{
Generate numeric XPath expression
}

