% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_and_checks.R
\name{rm_empty_cols}
\alias{rm_empty_cols}
\title{Remove columns which do not have data values}
\usage{
rm_empty_cols(df, header)
}
\arguments{
\item{df}{a data frame}

\item{header}{the header vector}
}
\value{
a data frame
}
\description{
Remove columns which do not have data values
}
\seealso{
\code{\link{rm_nuisance}, \link{rm_empty_rows}}
}

