% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_and_checks.R
\name{rm_empty_rows}
\alias{rm_empty_rows}
\title{Remove rows which do not have data values}
\usage{
rm_empty_rows(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame
}
\description{
Remove rows which do not have data values
}
\seealso{
\code{\link{rm_nuisance}, \link{rm_empty_cols}}
}

