% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_and_checks.R
\name{select_tab}
\alias{select_tab}
\title{Selects the table from the HTML Code}
\usage{
select_tab(which, Node)
}
\arguments{
\item{which}{a vector of length one for identification of the table in the document. Either
a numeric vector for the tables' rank or a character vector that describes an XPath for the table}

\item{Node}{the table node}

\item{...}{additional arguments passed to htmlParse}
}
\value{
a table node
}
\description{
Selects the table from the HTML Code
}

