% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/span_body.R
\name{span_body}
\alias{span_body}
\title{Creates body using span information}
\usage{
span_body(vals, colspans, rowspans)
}
\arguments{
\item{vals}{a list of body cell values}

\item{colspans}{a list of body colspans}

\item{rowspans}{a list of body rowspans}
}
\value{
a matrix of the assembled body
}
\description{
Creates body using span information
}

