% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/span_header.R
\name{span_header}
\alias{span_header}
\title{Creates header using span information}
\usage{
span_header(header.names, header.colspans, header.rowspans, headerSep)
}
\arguments{
\item{header.names}{a list of header names}

\item{header.colspans}{a list of header colspans}

\item{header.rowspans}{a list of header rowspans}

\item{headerSep}{a character vector that is used as a seperator in the construction of the table's variable names (default value ' >> ')}
}
\value{
a vector of header column names
}
\description{
Creates header using span information
}

