% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_colnames.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Produces correct column names}
\usage{
clean_colnames(header.name.table = NULL, colNames = NULL)
}
\arguments{
\item{header.name.table}{the header column names vector}

\item{colNames}{either a self-specificed character vector for the column names or a function used on header.name.table}
}
\value{
a character vector of header column names
}
\description{
Produces correct column names
}

