% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_rows.R
\name{get_body_xpath}
\alias{get_body_xpath}
\title{Return body xpath}
\usage{
get_body_xpath(body, table.Node)
}
\arguments{
\item{body}{an information for the body rows}

\item{table.Node}{the table node}
}
\value{
a character vector of XPath statements
}
\description{
Return body xpath
}

