% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_table_info.R
\name{get_cell_element}
\alias{get_cell_element}
\title{Extracts cells elements}
\usage{
get_cell_element(cells, tag = "td | th", elFun, rm_escape = NULL)
}
\arguments{
\item{cells}{a list of cell nodes}

\item{tag}{a character vector that provides information used in the XPath expression to extract the correct elements}

\item{elFun}{a function that is executed over the body cell nodes}

\item{rm_escape}{a character vector that, if specified, is used to replace escape sequences in header and body cells (default value ' ')}
}
\value{
the body element
}
\description{
Extracts cells elements
}

