% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{get_cell_element}
\alias{get_cell_element}
\title{Extracts cells elements}
\usage{
get_cell_element(cells, tag = "td | th", elFun, rm_escape, rm_whitespace)
}
\arguments{
\item{cells}{a list of cell nodes}

\item{tag}{a character vector that provides information used in the XPath expression to extract the correct elements}

\item{elFun}{a function that is executed over the header/body cell nodes}

\item{rm_escape}{a character vector that, if specified, is used to replace escape sequences in header and body cells (default value ' ')}

\item{rm_whitespace}{logical, should leading/trailing whitespace be removed from cell values (
default value TRUE)?}
}
\value{
the body element
}
\description{
Extracts cells elements
}

