% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cell_xpath}
\alias{get_cell_xpath}
\title{Construct an XPath expression for the body cells}
\usage{
get_cell_xpath(Node, body)
}
\arguments{
\item{Node}{the HTML document which can be a file name or a URL or an already parsed HTMLInternalDocument, or an HTML node of class XMLInternalElementNode, or a character vector containing the HTML content to parse and process}

\item{body}{a cha}
}
\value{
logical
}
\description{
Construct an XPath expression for the body cells
}
\examples{
add(1, 1)
add(10, 1)
}

