% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cell_xpath}
\alias{get_cell_xpath}
\title{Construct an XPath expression for the body cells}
\usage{
get_cell_xpath(table.Node, body)
}
\arguments{
\item{table.Node}{the table node}

\item{body}{body identifying information}
}
\value{
a character vector wth XPath statement
}
\description{
Construct an XPath expression for the body cells
}

