% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_cells}
\alias{get_cells}
\title{Extract body cell values}
\usage{
get_cells(Node, body)
}
\arguments{
\item{Node}{the HTML document which can be a file name or a URL or an already parsed HTMLInternalDocument, or an HTML node of class XMLInternalElementNode, or a character vector containing the HTML content to parse and process}

\item{body}{a cha}
}
\value{
logical
}
\description{
Extract body cell values
}
\examples{
add(1, 1)
add(10, 1)
}

