% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_cells.R
\name{get_cells}
\alias{get_cells}
\title{Extract body cell values}
\usage{
get_cells(table.Node, body)
}
\arguments{
\item{table.Node}{the table node}

\item{body}{body identifying information}
}
\value{
list of body information
}
\description{
Extract body cell values
}

