% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_head}
\alias{get_head}
\title{Assemble Check input values}
\usage{
get_head(table.Node, header)
}
\arguments{
\item{table.Node}{the table node}

\item{header}{an information on the header rows}
}
\value{
A list of header information
}
\description{
Assemble Check input values
}

