% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_head.R
\name{get_head}
\alias{get_head}
\title{Retrieve table head rows}
\usage{
get_head(table.Node, header)
}
\arguments{
\item{table.Node}{the table node}

\item{header}{a vector that contains information for the identification of the header row(s). A numeric vector can be specified where each element corresponds to the table rows. A character vector may be specified that describes an XPath for the header rows. If left unspecified, htmltable tries to use semantic information from the HTML code}
}
\value{
the header element
}
\description{
Retrieve table head rows
}

