% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_rows.R
\name{get_head_xpath}
\alias{get_head_xpath}
\title{Return header xpath}
\usage{
get_head_xpath(header, table.Node)
}
\arguments{
\item{header}{an information for the header rows}

\item{table.Node}{the table node}
}
\value{
a character vector of XPath statements
}
\description{
Return header xpath
}

