% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{get_header_elements}
\alias{get_header_elements}
\title{Extracts header elements}
\usage{
get_header_elements(cells, tag = "td | th")
}
\arguments{
\item{cells}{a list of cell nodes}

\item{tag}{a character vector that provides information used in the XPath expression to extract the correct elements}
}
\value{
A list of header information from the cells
}
\description{
Extracts header elements
}

