% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_xpath.R
\name{get_xpath}
\alias{get_xpath}
\title{Assemble XPath expressions for header and body}
\usage{
get_xpath(table.Node, header, body)
}
\arguments{
\item{table.Node}{the table node}

\item{header}{a vector that contains information for the identification of the header row(s). A numeric vector can be specified where each element corresponds to the table rows. A character vector may be specified that describes an XPath for the header rows. If left unspecified, htmltable tries to use semantic information from the HTML code}

\item{body}{a vector that specifies which table rows should be used as body information. A numeric vector can be specified where each element corresponds to a table row. A character vector may be specified that describes an XPath for the body rows. If left unspecified, htmltable tries to use semantic information from the HTML code}
}
\value{
a character vector of XPath statements
}
\description{
Assemble XPath expressions for header and body
}

