% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has_tag}
\alias{has_tag}
\title{Assert a specific tag in an XML node}
\usage{
has_tag(Node, tag)
}
\arguments{
\item{Node}{the HTML document which can be a file name or a URL or an already parsed HTMLInternalDocument, or an HTML node of class XMLInternalElementNode, or a character vector containing the HTML content to parse and process}

\item{tag}{a character string for the tag name to be matched}
}
\value{
logical
}
\description{
Assert a specific tag in an XML node
}
\examples{
add(1, 1)
add(10, 1)
}

