% Generated by roxygen2 (4.0.2): do not edit by hand
\name{htmltable}
\alias{htmltable}
\title{Assemble information from HTML tables}
\usage{
htmltable(doc, which = NULL, header = NULL, headerSep = " >> ",
  body = NULL, elFun = function(node) XML::xmlValue(node),
  colClasses = NULL, as.data.frame = TRUE, ...)
}
\arguments{
\item{doc}{the HTML document which can be a file name or a URL or an already parsed HTMLInternalDocument, or an HTML node of class XMLInternalElementNode, or a character vector containing the HTML content to parse and process.}

\item{which}{a vector identifying which tables to return from within the document (when such was specified). Either a numeric vector for the tables' rank or a character vector specifiying an XPath for the tables.}

\item{header}{a vector specifying header information. A numeric vector can be specified where each element corresponds to the table rows. A character vector may be specified that describes an XPath for the header rows.}

\item{headerSep}{a character vector that is used for constructing R table variable names from the HTML table.}

\item{body}{a vector specifying body information. A numeric vector can be specified where each element corresponds to the table rows. A character vector may be specified that describes an XPath for the body rows.}

\item{colClasses}{a vector of classes for the columns.}

\item{...}{additional arguments}

\item{cellFun}{a list of functions that is executed over the header and body cell nodes.}
}
\value{
A list of R data frames (if as.data.frame is left at its default value).
}
\description{
Assemble information from HTML tables
}
\examples{
This table lacks header information
We pass NULL to skip the header generation step
url <- "http://apps.who.int/immunization_monitoring/globalsummary/timeseries/tswucoveragebcg.html"
htmltable(doc = url, which = "/html/body/div/table[3]", header = NULL)


This example illustrates extracting multiple tables
url <-
}

