% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HTMLtable}
\alias{HTMLtable}
\title{Assemble information from HTML tables.}
\usage{
HTMLtable(table.node, which = NULL, tableSplit = NULL, header = NULL,
  headerSep = c(" >1> ", " >2> ", " >3> ", " >4> "), body = NULL,
  elFun = function(node) xmlValue(node), colClasses, as.data.frame = TRUE,
  ...)
}
\arguments{
\item{table.node}{A or character.}

\item{which}{Either numeric or character.}

\item{header}{numeric or character.}

\item{headerSep}{a vector}

\item{body}{numeric or character}

\item{colClasses}{a vector of classes}

\item{...}{additional arguments}

\item{cellFun}{A function on the node}
}
\value{
A list The sum of \code{x} and \code{y}.
}
\description{
Assemble information from HTML tables.
}
\examples{
add(1, 1)
add(10, 1)
}

