% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_xpath.R
\name{num_xpath}
\alias{num_xpath}
\title{Produce XPath for numeric index}
\usage{
num_xpath(index)
}
\arguments{
\item{index}{numeric vector}
}
\value{
a character vector of XPath statements
}
\description{
Produce XPath for numeric index
}

