% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rm_nuisance.R
\name{rm_empty_cols}
\alias{rm_empty_cols}
\title{Remove no data columns}
\usage{
rm_empty_cols(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a data frame
}
\description{
Remove no data columns
}

