% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rm_nuisance.R
\name{rm_nuisance}
\alias{rm_nuisance}
\title{Assemble Check input values}
\usage{
rm_nuisance(table.Node, rm_superscript = T, rm_footnotes = T)
}
\arguments{
\item{table.Node}{the table node}

\item{rm_superscript}{logical, denotes whether superscript information should be removed from header and body cells (default value TRUE)}

\item{rm_footnotes}{logical, denotes whether semantic footer information should be removed (default value TRUE)}
}
\value{
A list of table nodes.
}
\description{
Assemble Check input values
}

