% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_and_checks.R
\name{rm_nuisance}
\alias{rm_nuisance}
\title{Remove nuisance elements from the the table code}
\usage{
rm_nuisance(table.Node, rm_superscript, rm_footnotes, rm_invisible)
}
\arguments{
\item{table.Node}{the table node}

\item{rm_superscript}{logical, denotes whether superscript information should be removed from header and body cells (default value TRUE)}

\item{rm_footnotes}{logical, denotes whether semantic footer information should be removed (default value TRUE)}

\item{rm_invisible}{logical, should nodes that are not visible (display:none attribute) be removed?}
}
\value{
The revised table node
}
\description{
Remove nuisance elements from the the table code
}
\seealso{
\code{\link{rm_empty_cols}}
}

